Design: 스펙 리스트 아이템을 컴포넌트로 분리
